Moving runner to action.runner, moving tests to operator actions.
