+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/EventActions.scala
@@ -19,7 +19,7 @@ import java.time.ZonedDateTime
 
 import cloudflow.blueprint.deployment.StreamletDeployment
 import cloudflow.operator.action.EventActions.EventType.EventType
-import cloudflow.operator.runner.{ AkkaRunner, FlinkRunner, SparkRunner }
+import cloudflow.operator.action.runner.{ AkkaRunner, FlinkRunner, SparkRunner }
 import cloudflow.operator.{ CloudflowApplication, CloudflowLabels, DeploymentContext }
 import skuber.json.format.eventFmt
 import skuber.{ Event, ObjectEditor, ObjectMeta, ObjectResource }
