+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/FlinkRunnerActions.scala
@@ -21,7 +21,7 @@ import scala.collection.immutable._
 
 import skuber._
 
-import cloudflow.operator.runner._
+import cloudflow.operator.action.runner._
 
 object FlinkRunnerActions extends RunnerActions(FlinkRunner) {
   def apply(
