+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/PrepareNamespaceActions.scala
@@ -19,7 +19,7 @@ package action
 
 import scala.collection.immutable._
 import skuber._
-import cloudflow.operator.runner.{ AkkaRunner, FlinkRunner, Runner, SparkRunner }
+import cloudflow.operator.action.runner.{ AkkaRunner, FlinkRunner, Runner, SparkRunner }
 
 /**
  * Creates a sequence of resource actions for preparing the namespace where the application is
