+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/RunnerActions.scala
@@ -23,7 +23,7 @@ import org.slf4j.LoggerFactory
 import skuber.{ ObjectResource, Secret }
 import skuber.json.format._
 
-import cloudflow.operator.runner._
+import cloudflow.operator.action.runner._
 
 /**
  * Creates a sequence of resource actions for the runner changes
