+++ core/cloudflow-operator-actions/src/test/resources/serialization/app-blueprint-updated-v1-default
