+++ core/cloudflow-operator-actions/src/test/resources/serialization/app-deployment-accepted-v1-default
