+++ core/cloudflow-operator-actions/src/test/resources/serialization/app-deployment-failed-v1-default
