+++ core/cloudflow-operator-actions/src/test/resources/serialization/app-updated-v1-default
