+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/ResourceNamesSpec.scala
