+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/runner/FlinkRunnerSpec.scala
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
-package runner
+package cloudflow.operator.action.runner
 
 import com.typesafe.config.ConfigFactory
 import org.scalatest._
