+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/runner/PodConfigsSpec.scala
@@ -30,8 +30,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
-package runner
+package cloudflow.operator.action.runner
 
 import com.typesafe.config._
 import org.scalatest._
